refactor(header): split header component into a single file
